tdf#127410 Reliably copy doc print settings for MM
Commit 0f06fa5b9fe6192317f34dcd454032d790ad7244
("tdf#117664: mailmerge: copy printer settings to new mailmerge
document") already implemented copying print settings.

Move it into 'lcl_CreateWorkingDocument' so this happens
whenever a working document is created from the source
document.

The problem previously was that a printer settings weren't
previously copied to the working copy for the

    if( pSourceDocSh->IsModified() )

case in 'SwDBManager::Merge'.

Change-Id: I0dbe830bed0a36eb25da0d1b43a7f0d7dbc18f2e
Reviewed-on: https://gerrit.libreoffice.org/78714
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>