tdf#105330 sw: fix lost cursor on undoing nested table insert
This is a regression from commit
232a3ce5eba22c40469bae4e3ee62f969fd20aee (n#751313 SwCallLink: avoid
redrawing complete rows without nested tables, 2012-04-20), though
manual testing shows that the underlying problem has been addressed in
the meantime, so this can be reverted.

Over time, some poor tests started to depend on the new behavior so
adapt them as necessary:

1) Change back test added in commit
1981309bc42836c21652215b4b0fc0d984369cfc (testcase for fdo#38414,
2014-03-18) to its original form, that was changed to an export test in
commit 916e9a24d31d1888a40b96cd6a92c0f26c0a48d7 (move round-tripables to
ooxmlexport, 2016-10-07), as the export of that document still results
in data loss of cell content, just happened to pass so far.

2) Explicitly calculate content of text frames in two more tests, which
just hoped that by the time they assert, the layout is ready already
(but now that the missing notification is restored, it happens that the
first pass of the layout doesn't create them; only a later pass, invoked
by Idle, which doesn't run during cppunit tests).

Change-Id: Ie13d17266cd552f3850848ae980bcfe8776997ac
Reviewed-on: https://gerrit.libreoffice.org/79026
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>