tdf#104167 DOCX import: don't map Heading1 to Heading 1 para style
This is nominally a regression from commit
ba9237c3b87b5d5e503a6c3938c5515152801939 (writerfilter: reset properties
of reused styles, 2014-06-04), which assumed that style names are
unique.

They are not unique for the bugdoc, as it has paragraph styles named
'heading 1' and 'Heading1', both mapped to 'Heading 1'.

The mapping for the style name without a space was added in commit
1d2af74e226bf44e1a0f1e1981ca965cffeaa0b2 (INTEGRATION: CWS xmlfilter02
(1.27.12); FILE MERGED, 2008-01-10), in the "2007/11/22 14:53:43 os
1.27.12.9: Tab stops almost perfect now" part, sadly without any
indication on why that would be a good idea.

The binary DOC import never mapped Heading1 to Heading 1, so it doesn't
have this problem.

Fix the problem by removing this mapping, it leads to formatting loss
when a document has both HeadingN and Heading N styles.

Change-Id: I26dd999f5a536e079ac351ba5d40d20c83b99310
Reviewed-on: https://gerrit.libreoffice.org/79422
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>