tdf#107784: DOCX Import: Show citation's title in fields (2nd take)
Regression from 452a35dc59f6f56a930165bbf78bb227f9a2c0a1

The problem as I see it is that LibreOffice uses the identifier
plus optionals prefix and suffix to display the bibliographic
references.
On the other hand, MSO displays the title for references
starting with CITATION.

So do the same MSO does, when working with CITATION refs

We already add CITATION when we export to ooxml.
See case SwFieldIds::TableOfAuthorities in
AttributeOutputBase::TextField

Change-Id: I144f27f711926658c8b6f89e69f0ddeeb7e7890c
Reviewed-on: https://gerrit.libreoffice.org/79477
Tested-by: Jenkins
Reviewed-by: Xisco Faulí <xiscofauli@libreoffice.org>