tdf#127701 Zoom factor percentage calculating fault
regression from
    commit 7e9842d66b537de03690d9e2efa6c7d5d886d1ae
tdf#126819 Can't retrieve the mouse position written in the
statusbar

note that this is a workaround - we fire an sfx2 event when we choose
one of the "smart" zoom options, and normally we then receive another
sfx2 event to notify us of the real zoom value.
But when we choose the same smart option twice, we don't receive the
second sfx2 event.

Change-Id: I31c2e34ff31a4f4e03a447ffbd783589982edc57
Reviewed-on: https://gerrit.libreoffice.org/79516
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>