tdf#116685: Make the RICHTEXT take precedence over EMBED_SOURCE.
Before this patch, copy in Calc, Paste in Writer produced an embedded
sheet.  I suspect is it not what the people usually want; working with
the embedded sheets in Writer is non-intuitive, I suspect people will be
happier with a normal table which they can style etc. appropriately.

OTOH - this is a general change, so it might have some unwanted
side-effects; let's see what if we get bugreports :-)

tdf#127673 was related; but in my view, we shouldn't create the embedded
objects in the first place.

Change-Id: I73710168d8924fdd6275d65e20cf55f38971a3f9
Reviewed-on: https://gerrit.libreoffice.org/79606
Tested-by: Jenkins
Reviewed-by: Jan Holesovsky <kendy@collabora.com>