uitest: make child soffice process die when parent dies
using PR_SET_PDEATHSIG.

Note that this can only be called by the child process in question, so
we have to pass a new command line argument down to activate it.

And we have to plumb it through the wrapper oosplash process.

I wonder where else our testing infrastructure could benefit from this

Change-Id: I55e8e9f7f4e6cc415046df04c804e51475b8a4c9
Reviewed-on: https://gerrit.libreoffice.org/79854
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>