tdf#127446 vcl image lazy-load: fix custom size handling of metafiles
This is a regression from commit
fdceee1c686ce64357583c9869c6a4d71cb2bebd (tdf#125591 DOC import:
lazy-load metafiles with explicit size, 2019-06-11), which assumed that
once maSwapInfo.maPrefSize stores the preferred size, it'll be set on
the Graphic when it's loaded later.

It seems there was no support for that, it was just an accident that the
guessed size of the metafile was about right.

Handle this explicitly in ImpGraphic::loadPrepared(), so that the bugdoc
(which has a custom preferred size) aspect ratio is correct.

Change-Id: Ic7c4009ad6723a2e16129d27600c904311ff3daf
Reviewed-on: https://gerrit.libreoffice.org/79899
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>