tdf#124601 sw FollowTextFlow: fix hori pos of objects outside the current cell
The problem is that the image in the bugdoc's footer is anchored in one
cell, but its position has a value that shifts it to the next column and
next row. The next column is the problem for the horizontal position.

So build on top of the previous vertical position fix, and make sure
that CalcRelPosX() doesn't limit the position inside the current cell
for an in-table, follow-text-flow, wrap-though image.

Once that's in place, make sure that we don't try to grow the cell due
to follow-text-flow, wrap-though objects in CalcPosition(), since
in a wrap-through vs follow-text-flow situation, the wrap-though should
have priority (should not affect size of cells).

Finally, now that cells don't grow in this case, the previously added
special-casing of footers in SwFlyFreeFrame::CheckClip() is no longer
necessary. Not growing the cells means we don't try to re-position the
object.

Change-Id: Ic55e4b5188704fa70314f91fe9a01987b6a56d7b
Reviewed-on: https://gerrit.libreoffice.org/80032
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins