tdf#126589: only consider text to the left of cursor as AutoText short name
Given a text "This is a dtfoo bar", and cursor inside "dtfoo" between "t"
and "f", invoking AutoText function (F3) currently considers the whole word
"dtfoo" as autotext short name. This changes it to only consider the part of
the word to the left of cursor, i.e. only "dt" in the example. This removes
the requirement to have a word boundary after the autotext short name.

Change-Id: I6ba28e63fe25664131b7e03dda39ae997dffa464
Reviewed-on: https://gerrit.libreoffice.org/80419
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>