tdf#118947 sw tablestyle: manually scan parents for ::SET
...since we do not want to find DocDefaults or pool defaults.

This patch mainly does two things.
1.) Since DocDefaults get merged into root-level styles,
this patch identifies these DocDefaults as values that
can be overwritten by table style properties.

2.) Some properties (like left/right margins) are paired
together, so if one is SET, the other is also initialized
and reported as SET. These are identified, and
followed to see if they actually inherit from pool defaults.

Change-Id: I4291c77073858ae23360a9f34d3650bfd5df86ca
Reviewed-on: https://gerrit.libreoffice.org/81413
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>