sw: add CH_TXT_ATR_FIELDSEP position parameter when inserting
Different callers want it at the end, at the start, or in the case of
Undo it can even be in an arbitrary position... (except if it is ensured
that the field command is deleted before the fieldmark?).

So let's just pass in SwPosition and be flexible, if it is missing it
will be before the end as before.

Change-Id: Ibec222f633bdaf66abd1540027d0f5c75988c738
Reviewed-on: https://gerrit.libreoffice.org/80673
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>