tdf#127964 PPTX import: fix shape fill handling: style vs slide background
Regression from commit ff0b4d23f19eca406271a230d1a4c6079517579c
(tdf#123684 PPTX import: fix wrong background color for <p:sp
useBgFill="1">, 2019-04-23), the problem was that we didn't handle the
case when a shape had an XML fragment like this:

      <p:sp useBgFill="1">
        <p:style>
          <a:fillRef idx="1">
            <a:schemeClr val="accent1"/>
          </a:fillRef>
        </p:style>
      </p:sp>

i.e. the shape both wants to use background fill and it has a style
declaring how to fill it as well. We gave the style a priority, while
PowerPoint gives the background fill a priority.

Fix the problem by not setting the fill from the style in case the
background fill is already set.

Change-Id: Ie1b56e5615219138a5b7ddd7a2b25295b991bc05
Reviewed-on: https://gerrit.libreoffice.org/80804
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>