tdf#86188 Compatibility setting enabling comments to footnotes
For better OOXML interoperability allow disabling the insertion
of comments to footnotes/endnotes.

It is not possible to add comments to footnotes/endnotes in Word
so no wonder Writer generates invalid XML if one tries to save
comments added to footnotes as DOCX.

Prevent that by adding a centrally manageably key for disabling the
menu item if the cursor is in a footnote/endnote. In OOXML-heavy
environments sysadmins should be able to disable this and thus
hide the footgun from users.

Default setting is true so it means no change compared to current
feature set.

Change-Id: I2f799cb3f77a47fc14fa60b55fc5689a2710aff7
Reviewed-on: https://gerrit.libreoffice.org/80829
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>