tdf#120224 writerfilter: don't suppress all bookmarks inside fields
MSWordExportBase used to erroneously export all sorts of marks as
bookmarks, but we fixed that.  Apparently this code in
DomainMapper_Impl::SetBookmarkName() should prevent import of the
bookmarks that were exported as duplicates for fieldmarks; suppress
only the marks with the same name as the fieldmark, but not unrelated
marks that happen to start inside the field, like the "_GoBack" one in
tdf120224_textControlCrossRef.docx.

The reason why this was necessary previously is some dubious
nonsense in model.xml that caused the ffData stuff to be delayed until
the next run, hence the bookmark start is sent to domain mapper before
the ffData; this was already helpfully marked with a TODO
comment, so just implement the suggestion.

(regression from ccadb3eb11d2ba380e47dcad6705de3ccd423a37)

Change-Id: I15a1865ac34b0b9b3f11849d612e95115582b90b
Reviewed-on: https://gerrit.libreoffice.org/80917
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>