lok: get spelling context menu on long press
This patch handles a new flag attached to the invalidate view cursor
message for informing the client when the text cursor is inside a
mispelled word.
This information is used for popping up the spelling context menu on a
long press event instead of the standard context menu for a selected
word.

Change-Id: I13fcbe53c83ca6eb56300a601734cdc3211e88a0
Reviewed-on: https://gerrit.libreoffice.org/85244
Tested-by: Jenkins
Reviewed-by: Tamás Zolnai <tamas.zolnai@collabora.com>