We do not need a C++ GNU dialect
a71c16a5cf1f0105dd521f2e410a7cb7518eb23e "use -std=gnu++0x rather than
-std=c++0x" appears to have started this, but the only rationale it gives is
that it keeps things in sync with GCC's default behavior when no -std= is given.
But it apparently works fine to build with a -std=c++... standard dialect.

This allows to get rid of the check introduced with
2b6424e0676f9a068edd30ee1ae1cedbb65b12ed "Ensure configuration that defines
math_errhandling in <cmath>".

(It kept bothering me to say "I observe this-and-this with -std=c++2a" when what
configure.ac made me actually use was -std=gnu++2a.  And truthfully saying
"-std=gnu++2a" would have been a distraction, as what is relevant for such an
observation is most likely the "2a" and not the "gnu" part.)

Change-Id: I7c213a702ffb7df6f4c2c4a421008e30e2712a51
Reviewed-on: https://gerrit.libreoffice.org/81176
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>