tdf#128230 Require password if the Unprotect button is visible
Regression from commit 04456434798935a9dbb3a328df72e70bd4d3214d

Before that requiring password depended on the current text of
the Protect/Unprotect button, which was changed dynamically.

This commit changed that logic and introduced two buttons in the
new .ui file that are hidden/shown dynamically.

The password condition however was changed to check the visibility
of the new Protect button instead of the Unprotect button.

Change-Id: Ie24e1b2d45fa92a375a29b7bc71689f9b83ae9dc
Reviewed-on: https://gerrit.libreoffice.org/81283
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>