tdf#128460 import SwInputField from DOCX FILLIN
Actually the whole FILLIN handling is not that correct, because
the value can actually be formatted in any way in Word, as it's
just defined as a range of runs with formating, while Writer just
applies the character formatitng to the whole field.

But this at least allows for a round trip of the exported Writer
type input field after commit 29ce9d780eba ("tdf#125103 Writer:
input fields will be exported to docx") changed this.

Change-Id: I6389b4a2ff9d2b62e1e4a00f1c1e112c199377b5
Regression-from: 29ce9d780eba389f5e902b7f6d6a57886a00028d
Reviewed-on: https://gerrit.libreoffice.org/81686
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>