Fix IsTransparent() for unloaded graphics
Fix Graphic::IsAlpha() and Graphic::IsTransparent() for unloaded
graphics. This fixes tdf#118036.

GraphicDescriptor::Detect(true) is currently used to read the image size
from the header of images which are not being fully loaded yet.  This
change extends GraphicDescriptor to also report whether the image
supports transparency or alpha, implemented only for PNG format so far.

Change-Id: I1753c0d11491f1dc518e23da8d7b3842945770cb
Reviewed-on: https://gerrit.libreoffice.org/81785
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>