Workaround for crash on macOS Catalina at least when built with Xcode 11
For some unclear reason, for me at least, when I have built with Xcode
11.1 or 11.2 and run on macOS Cataline, I run into a crash caused by
boundless recursion related to the progress bar in the recovery
dialog. Not setting mbProgressNeedsErase to true for the "NWF" seems
to help.

Whether this has some unintended visually bad side effect I don't
know, but hey, it's just a progress bar, who cares if it in some
circumstances doesn't look perfect?

Change-Id: I6c990adde7689633b2c4b5aa42a4a07966565af2
Reviewed-on: https://gerrit.libreoffice.org/81874
Tested-by: Jenkins
Reviewed-by: Tor Lillqvist <tml@collabora.com>
Tested-by: Tor Lillqvist <tml@collabora.com>