tdf#128335 sw textbox: fix crash on deleting multiple shapes with textboxes
This was introduced in commit c78b3f5d8b50668054683ca21439d5283b80c33c
(SwDrawView::DeleteMarked: delete textbox of shape as well, 2014-06-23),
the problem is that in case one shape+textframe pair is anchored in an
other textframe, which is also part of a shape+textframe pair, it
matters what order we use for deleting these objects.

What happened is that the anchor position of a 2nd textbox was updated
on the deletion (cut) of the 1st textbox, and this was not properly
restored on paste.

Fix the problem by deleting in reverse order, assuming that the
container is ordered by anchor positions.

Change-Id: I62bc12a5064ce397234999e8fd7b108f79de3c24
Reviewed-on: https://gerrit.libreoffice.org/82042
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>